[#16, contentAssist] add support for custom parser entry points
ContentAssist needs to use the entry point that can be obtained via 

NodeModelUtils.getEntryParserRule(INode) or 
XtextResource.getEntryPoint()

To support this, AbstractContentAssistParser does now implement 
IPartialEditingContentAssistParser, which is an interface that has been
introduced some time ago for the CodeTemplates language.

XtextAntlrGeneratorFragment2 will no longer generate getFollowElements()
for each language because that method is only able to invoke the first 
parser rule. Instead, there is now a gneric implementation in the super 
class.

Signed-off-by: Moritz Eysholdt <moritz.eysholdt@typefox.io>